Rearrange some UI
Move upload button to home screen
Add some padding to all screens so ui isn't shoved to the top
