+++ composeApp/src/androidMain/kotlin/com/ramitsuri/locationtracking/ui/home/HomeScreen.kt
@@ -8,11 +8,15 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.filled.Upload
 import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -32,6 +36,7 @@ fun HomeScreen(
     onNavToSettings: () -> Unit,
     onNavToAppSettings: () -> Unit,
     onSingleLocation: () -> Unit,
+    onUpload: () -> Unit,
 ) {
     Column(
         modifier = modifier
@@ -42,6 +47,16 @@ fun HomeScreen(
             modifier = Modifier.fillMaxWidth(),
             horizontalArrangement = Arrangement.End,
         ) {
+            IconButton(onClick = onUpload) {
+                if (viewState.isUploadWorkerRunning) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(20.dp),
+                        color = MaterialTheme.colorScheme.onBackground,
+                    )
+                } else {
+                    Icon(Icons.Default.Upload, null)
+                }
+            }
             IconButton(onClick = onNavToSettings) {
                 Icon(Icons.Default.Settings, null)
             }
@@ -56,6 +71,7 @@ fun HomeScreen(
         Spacer(modifier = Modifier.height(16.dp))
         Button(
             onClick = onSingleLocation,
+            modifier = Modifier.fillMaxWidth(),
         ) {
             Text(stringResource(R.string.single_location))
         }
