+++ composeApp/src/commonMain/kotlin/com/ramitsuri/locationtracking/ui/home/HomeViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.ramitsuri.locationtracking.permissions.PermissionResult
 import com.ramitsuri.locationtracking.repository.LocationRepository
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.combine
@@ -13,6 +14,8 @@ import kotlinx.coroutines.flow.stateIn
 class HomeViewModel(
     locationRepository: LocationRepository,
     permissionState: () -> StateFlow<List<PermissionResult>>,
+    isUploadWorkerRunning: () -> Flow<Boolean>,
+    private val upload: () -> Unit,
 ) : ViewModel() {
     val viewState = combine(
         locationRepository.getCount(),
@@ -21,15 +24,20 @@ class HomeViewModel(
                 !permissionResult.granted
             }.map { it.permission }
         },
-    ) { count, permissions ->
+        isUploadWorkerRunning(),
+    ) { count, permissions, isUploadWorkerRunning ->
         HomeViewState(
             numOfLocations = count,
             notGrantedPermissions = permissions,
+            isUploadWorkerRunning = isUploadWorkerRunning,
         )
+    }.stateIn(
+        scope = viewModelScope,
+        started = SharingStarted.WhileSubscribed(5_000),
+        initialValue = HomeViewState(),
+    )
+
+    fun onUpload() {
+        upload()
     }
-        .stateIn(
-            scope = viewModelScope,
-            started = SharingStarted.WhileSubscribed(5_000),
-            initialValue = HomeViewState(),
-        )
 }
