+++ composeApp/src/commonMain/kotlin/com/ramitsuri/locationtracking/ui/home/HomeViewState.kt
@@ -5,4 +5,5 @@ import com.ramitsuri.locationtracking.permissions.Permission
 data class HomeViewState(
     val numOfLocations: Int = 0,
     val notGrantedPermissions: List<Permission> = emptyList(),
+    val isUploadWorkerRunning: Boolean = false,
 )
